Making things more pythonic, but now noticing there's an issue with the patches notebook
