--- dreimac/Utils.py
@@ -1,109 +1,0 @@
-import numpy as np
-from scipy import sparse
-
-
-"""#########################################
-        Cohomology Utility Functions
-#########################################"""
-
-def add_cocycles(c1, c2, p = 2, real = False):
-    S = {}
-    c = np.concatenate((c1, c2), 0)
-    for k in range(c.shape[0]):
-        [i, j, v] = c[k, :]
-        i, j = min(i, j), max(i, j)
-        if not (i, j) in S:
-            S[(i, j)] = v
-        else:
-            S[(i, j)] += v
-    cret = np.zeros((len(S), 3))
-    cret[:, 0:2] = np.array([s for s in S])
-    cret[:, 2] = np.array([np.mod(S[s], p) for s in S])
-    dtype = np.int64
-    if real:
-        dtype = np.float32
-    cret = np.array(cret[cret[:, -1] > 0, :], dtype = dtype)
-    return cret
-
-def make_delta0(R):
-    """
-    Return the delta0 coboundary matrix
-    :param R: NEdges x 2 matrix specifying edges, where orientation
-    is taken from the first column to the second column
-    R specifies the "natural orientation" of the edges, with the
-    understanding that the ranking will be specified later
-    It is assumed that there is at least one edge incident
-    on every vertex
-    """
-    NVertices = int(np.max(R) + 1)
-    NEdges = R.shape[0]
-    
-    #Two entries per edge
-    I = np.zeros((NEdges, 2))
-    I[:, 0] = np.arange(NEdges)
-    I[:, 1] = np.arange(NEdges)
-    I = I.flatten()
-    
-    J = R[:, 0:2].flatten()
-    
-    V = np.zeros((NEdges, 2))
-    V[:, 0] = -1
-    V[:, 1] = 1
-    V = V.flatten()
-    I = np.array(I, dtype=int)
-    J = np.array(J, dtype=int)
-    Delta = sparse.coo_matrix((V, (I, J)), shape=(NEdges, NVertices)).tocsr()
-    return Delta
-
-
-"""#########################################
-        Partition of Unity Functions
-#########################################"""
-
-def partunity_linear(ds, r_cover):
-    """
-    Parameters
-    ----------
-    ds: ndarray(n)
-        Some subset of distances between landmarks and 
-        data points
-    r_cover: float
-        Covering radius
-    Returns
-    -------
-    varphi: ndarray(n)
-        The bump function
-    """
-    return r_cover - ds
-
-def partunity_quadratic(ds, r_cover):
-    """
-    Parameters
-    ----------
-    ds: ndarray(n)
-        Some subset of distances between landmarks and 
-        data points
-    r_cover: float
-        Covering radius
-    Returns
-    -------
-    varphi: ndarray(n)
-        The bump function
-    """
-    return (r_cover - ds)**2
-
-def partunity_exp(ds, r_cover):
-    """
-    Parameters
-    ----------
-    ds: ndarray(n)
-        Some subset of distances between landmarks and 
-        data points
-    r_cover: float
-        Covering radius
-    Returns
-    -------
-    varphi: ndarray(n)
-        The bump function
-    """
-    return np.exp(r_cover**2/(ds**2-r_cover**2))
